requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==46.0.3
cssselect2==0.8.0
Django==5.1.7
django-crispy-forms==2.3
et_xmlfile==2.0.0
freetype-py==2.5.1
html5lib==1.1
idna==3.11
lxml==6.0.2
openpyxl==3.1.5
oscrypto==1.3.0
pillow==11.2.1
pycairo==1.28.0
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pypdf==6.1.3
python-bidi==0.6.7
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
# --- Producción ---
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
